requirements.txt
pandas~=2.1.1
s3fs~=2023.10.0
configparser~=6.0.0
apache-airflow~=2.7.3
requests~=2.31.0